# openapi_server/models/task_agent_assignment.py
# coding: utf-8

"""
    AutoGen Framework API

    API for managing agents, teams, and tasks in an AutoGen system.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501
from datetime import datetime
from typing import Optional
from sqlmodel import SQLModel, Field

class TaskAgentAssignment(SQLModel, table=True):
    task_id: int = Field(foreign_key="task.task_id", primary_key=True)
    agent_id: int = Field(foreign_key="agent.agent_id", primary_key=True)
    role_in_task: str
    assigned_at: Optional[datetime] = None