# openapi_server/models/task_agent_assignment_input.py
# coding: utf-8

"""
    AutoGen Framework API

    API for managing agents, teams, and tasks in an AutoGen system.

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from typing import Optional
from sqlmodel import SQLModel, Field

class TaskAgentAssignmentInput(SQLModel):
    agent_id: int
    role_in_task: str